Refactor pkg names for 'revive' linter error
