+++ csc/cmd/controller_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/container-storage-interface/spec/lib/go/csi"
 	"github.com/dell/gocsi/mock/service"
-	"github.com/dell/gocsi/utils"
+	utils "github.com/dell/gocsi/utils/csi"
 	"github.com/spf13/cobra"
 	"github.com/stretchr/testify/assert"
 )
