+++ middleware/serialvolume/lockprovider/volume_lock_provider.go
@@ -1,4 +1,4 @@
-package types
+package lockprovider
 
 import (
 	"context"
