+++ mock/service/service.go
@@ -8,7 +8,7 @@ import (
 	"sync/atomic"
 
 	"github.com/container-storage-interface/spec/lib/go/csi"
-	"github.com/dell/gocsi/utils"
+	utils "github.com/dell/gocsi/utils/csi"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
