+++ testing/controller_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/container-storage-interface/spec/lib/go/csi"
 
 	"github.com/dell/gocsi/mock/service"
-	"github.com/dell/gocsi/utils"
+	utils "github.com/dell/gocsi/utils/csi"
 )
 
 var _ = Describe("Controller", func() {
