+++ utils/csi/csi.go
@@ -1,4 +1,4 @@
-package utils
+package csi
 
 import (
 	"bytes"
