+++ utils/middleware/middleware.go
@@ -1,4 +1,4 @@
-package utils
+package middleware
 
 import (
 	"reflect"
