+++ utils/rpcs/rpcs.go
@@ -1,4 +1,4 @@
-package utils
+package rpcs
 
 import (
 	"fmt"
